Sat, Feb 19, 2022, 12:13 AM +00:00

+++ caesar/caesar.c
@@ -62,7 +62,7 @@ char rotate(char c, int key)
     if (islower(c))
     {
 
-        printf("%c\n", (c+key)%26);
+        printf("%c\n", ((int)c+key)%26);
     }
 
     return c;
